leads.py
import json
import pandas as pd
import requests
import time
import logging
import os
from bs4 import BeautifulSoup
from datetime import datetime

# Configure logging
logging.basicConfig(level=logging.DEBUG)
logger = logging.getLogger(__name__)

def classify_domain(domain):
    """
    Classify a company domain into a business category
    """
    domain = (domain or '').lower()
    if 'tech' in domain or 'software' in domain or 'digital' in domain or 'comp' in domain or 'it' in domain:
        return 'Technology'
    elif 'finance' in domain or 'bank' in domain or 'invest' in domain or 'capital' in domain:
        return 'Finance'
    elif 'health' in domain or 'med' in domain or 'care' in domain or 'pharm' in domain:
        return 'Healthcare'
    elif 'edu' in domain or 'school' in domain or 'college' in domain or 'univ' in domain:
        return 'Education'
    elif 'marketing' in domain or 'media' in domain or 'advert' in domain or 'pr' in domain:
        return 'Marketing'
    elif 'retail' in domain or 'shop' in domain or 'store' in domain:
        return 'Retail'
    elif 'manufact' in domain or 'indust' in domain or 'prod' in domain:
        return 'Manufacturing'
    elif 'consult' in domain or 'service' in domain or 'solution' in domain:
        return 'Consulting'
    else:
        return 'Other'

def scrape_linkedin_company_page(url, user_agent='Mozilla/5.0', timeout=10):
    """
    Scrape a LinkedIn company page for company information
    """
    headers = {'User-Agent': user_agent}
    logger.debug(f"Scraping LinkedIn URL: {url}")
    
    try:
        resp = requests.get(url, headers=headers, timeout=timeout)
        if resp.status_code != 200:
            logger.warning(f"Failed to retrieve {url} - Status code: {resp.status_code}")
            return None
        
        soup = BeautifulSoup(resp.text, 'html.parser')
        
        # Extract company name
        name_elem = soup.find('h1')
        name = name_elem.text.strip() if name_elem else ''
        
        # Extract description
        desc_elem = soup.find('meta', {'name': 'description'})
        desc = desc_elem['content'].strip() if desc_elem and desc_elem.has_attr('content') else ''
        
        # Extract domain from URL
        domain = url.split('/')[4] if len(url.split('/')) > 4 else ''
        
        # Try to find website URL
        website = ''
        for a in soup.find_all('a', href=True):
            if 'website' in a.text.lower():
                website = a['href']
                break
        
        # Try to extract company size
        size = ''
        for p in soup.find_all('p'):
            if 'employees' in p.text.lower():
                size = p.text.strip()
                break
        
        # Try to extract location
        location = ''
        for span in soup.find_all('span'):
            if 'headquarter' in span.text.lower():
                location = span.text.strip()
                break
        
        # Try to extract founding year
        founded = ''
        for p in soup.find_all('p'):
            if 'founded' in p.text.lower():
                founded = p.text.strip()
                break
        
        return {
            'companyLinkedinUrl': url,
            'name': name,
            'description': desc,
            'website': website,
            'domain': domain,
            'domain_class': classify_domain(domain),
            'size': size,
            'location': location,
            'founded': founded,
            'scraped_at': datetime.now().strftime('%Y-%m-%d %H:%M:%S')
        }
    except Exception as e:
        logger.error(f"Error scraping {url}: {str(e)}")
        return None

def google_search_linkedin_companies(query, max_results=10):
    """
    Search for LinkedIn company URLs via Google
    """
    logger.debug(f"Searching for: {query}")
    urls = []
    
    try:
        # Try using googlesearch-python
        from googlesearch import search
        for url in search(query, num_results=max_results*2):  # Get more results to filter
            if 'linkedin.com/company/' in url and url not in urls:
                urls.append(url)
                if len(urls) >= max_results:
                    break
    except ImportError:
        logger.warning("googlesearch-python not installed, using alternative method")
        try:
            # Try using custom search with requests + BeautifulSoup
            search_url = f"https://www.google.com/search?q={query.replace(' ', '+')}"
            headers = {'User-Agent': 'Mozilla/5.0'}
            response = requests.get(search_url, headers=headers)
            
            if response.status_code == 200:
                soup = BeautifulSoup(response.text, 'html.parser')
                links = soup.find_all('a')
                
                for link in links:
                    href = link.get('href', '')
                    if 'linkedin.com/company/' in href and href not in urls:
                        # Extract actual URL from Google redirect
                        if '/url?q=' in href:
                            href = href.split('/url?q=')[1].split('&')[0]
                        urls.append(href)
                        if len(urls) >= max_results:
                            break
        except Exception as e:
            logger.error(f"Error in custom Google search: {str(e)}")
            # Fall back to mock data if all else fails
            urls = [
                "https://uk.linkedin.com/company/microsoft",
                "https://uk.linkedin.com/company/google",
                "https://uk.linkedin.com/company/amazon"
            ]
            
    logger.debug(f"Found {len(urls)} LinkedIn URLs")
    return urls[:max_results]  # Ensure we only return up to max_results

def run_scraper(
    keywords=None,
    founded_years=None,
    country=None,
    size=None,
    config_path='scraper_config.json',
    max_results=10,
    user_agent='Mozilla/5.0',
    timeout=10,
    output_csv='lead1.csv',
    sleep_time=1.0,
    search_func=None
):
    """
    Main function to run the LinkedIn company scraper
    """
    # Load config if it exists
    config = {}
    if os.path.exists(config_path):
        with open(config_path, 'r') as f:
            config = json.load(f)
    
    # Use provided parameters or defaults from config
    keywords = keywords or config.get('keywords', 'IT services')
    founded_years = founded_years or config.get('founded_years', ['2015'])
    country = country or config.get('country', 'United kingdom')
    size = size or config.get('size', '51-200')
    
    # Build search query
    query = f"{keywords} companies founded {','.join(founded_years)} in {country} with {size} employees linkedin"
    logger.info(f"Search Query: {query}")
    
    # Get LinkedIn company URLs
    urls = []
    if search_func is not None:
        urls = search_func(query)
    else:
        urls = google_search_linkedin_companies(query, max_results)
    
    # Scrape each company page
    results = []
    for i, url in enumerate(urls):
        logger.info(f"Scraping {i+1}/{len(urls)}: {url}")
        data = scrape_linkedin_company_page(url, user_agent=user_agent, timeout=timeout)
        if data:
            results.append(data)
        # Add sleep to avoid rate limiting
        if i < len(urls) - 1:  # Don't sleep after the last one
            logger.debug(f"Sleeping for {sleep_time} seconds")
            time.sleep(sleep_time)
    
    # Create DataFrame from results
    df = pd.DataFrame(results)
    
    # Save to CSV if output path provided
    if output_csv:
        df.to_csv(output_csv, index=False)
        logger.info(f"Results saved to {output_csv}")
    
    return df

if __name__ == '__main__':
    import argparse
    parser = argparse.ArgumentParser(description='LinkedIn Company Scraper')
    parser.add_argument('--keywords', type=str, default=None, help='Keywords for search')
    parser.add_argument('--founded_years', type=str, default=None, help='Comma-separated years')
    parser.add_argument('--country', type=str, default=None, help='Country')
    parser.add_argument('--size', type=str, default=None, help='Company size')
    parser.add_argument('--config_path', type=str, default='scraper_config.json', help='Config file path')
    parser.add_argument('--max_results', type=int, default=10, help='Max LinkedIn results')
    parser.add_argument('--user_agent', type=str, default='Mozilla/5.0', help='User agent')
    parser.add_argument('--timeout', type=int, default=10, help='Request timeout')
    parser.add_argument('--output_csv', type=str, default='lead1.csv', help='Output CSV file')
    parser.add_argument('--sleep_time', type=float, default=1.0, help='Sleep time between requests (seconds)')
    
    args = parser.parse_args()
    founded_years = args.founded_years.split(',') if args.founded_years else None
    
    df = run_scraper(
        keywords=args.keywords,
        founded_years=founded_years,
        country=args.country,
        size=args.size,
        config_path=args.config_path,
        max_results=args.max_results,
        user_agent=args.user_agent,
        timeout=args.timeout,
        output_csv=args.output_csv,
        sleep_time=args.sleep_time
    )
    
    print(f"Scraped {len(df)} companies and saved to {args.output_csv}")
