pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "googlesearch-python>=1.3.0",
    "gunicorn>=23.0.0",
    "nltk>=3.9.1",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "trafilatura>=2.0.0",
]
